+++ apps/dashboard/src/app/knowledge-box/knowledge-box-users/users-manage/users-manage.service.ts
@@ -1,6 +1,6 @@
 import { Injectable } from '@angular/core';
 import { BehaviorSubject, combineLatest, Observable } from 'rxjs';
-import { shareReplay, switchMap, tap } from 'rxjs/operators';
+import { map, shareReplay, switchMap, take, tap } from 'rxjs/operators';
 import { SDKService } from '@flaps/core';
 import { InviteKbData, KBRoles } from '@nuclia/core';
 
@@ -14,7 +14,18 @@ export class UsersManageService {
         switchMap(([kb, account]) => kb.getUsers(account.slug)),
       ),
     ),
-    shareReplay(),
+    shareReplay(1),
+  );
+
+  invitesKb = this._onUpdateUsers.pipe(
+    switchMap(() => this.sdk.currentKb),
+    switchMap((kb) => kb.getInvites()),
+    map((invites) =>
+      invites.map(
+        (invite) => ({ ...invite, expires: invite.expires + 'Z' }), // Set timezone
+      ),
+    ),
+    shareReplay(1),
   );
 
   constructor(private sdk: SDKService) {}
@@ -24,11 +35,15 @@ export class UsersManageService {
   }
 
   inviteUser(data: InviteKbData): Observable<void> {
-    return this.sdk.currentKb.pipe(switchMap((kb) => kb.inviteToKb(data)));
+    return this.sdk.currentKb.pipe(
+      take(1),
+      switchMap((kb) => kb.inviteToKb(data)),
+    );
   }
 
   changeRole(userId: string, role: KBRoles) {
     return this.sdk.currentKb.pipe(
+      take(1),
       switchMap((kb) => kb.updateUsers({ update: [{ id: userId, role }] })),
       tap(() => this.updateUsers()),
     );
@@ -36,7 +51,16 @@ export class UsersManageService {
 
   deleteUser(userId: string) {
     return this.sdk.currentKb.pipe(
+      take(1),
       switchMap((kb) => kb.updateUsers({ delete: [userId] })),
+      tap(() => this.updateUsers()),
+    );
+  }
+
+  deleteInvite(email: string) {
+    return this.sdk.currentKb.pipe(
+      take(1),
+      switchMap((kb) => kb.deleteInvite(email)),
       tap(() => this.updateUsers()),
     );
   }