+++ libs/user/src/lib/login/login.component.ts
@@ -2,7 +2,7 @@ import { Component, ElementRef, ViewChild } from '@angular/core';
 import { FormControl, FormGroup, Validators } from '@angular/forms';
 import { ActivatedRoute, Router } from '@angular/router';
 import { ReCaptchaV3Service } from 'ngx-captcha';
-import { catchError, map, of, switchMap } from 'rxjs';
+import { catchError, distinctUntilChanged, map, of, switchMap } from 'rxjs';
 
 import { BackendConfigurationService, OAuthService, SAMLService, SDKService } from '@flaps/core';
 import { InputComponent } from '@guillotinaweb/pastanaga-angular';
@@ -46,6 +46,7 @@ export class LoginComponent {
   isLoggingIn = false;
 
   ssoUrl = this.loginForm.controls.email.valueChanges.pipe(
+    distinctUntilChanged(),
     switchMap((email) => {
       const parts = (email || '').split('@');
       const domain = parts.length > 1 ? parts.pop() : undefined;