+++ libs/sdk-core/CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Support Knowledge Box notification endpoint
 - Add support for new `/processing-status` endpoint in `kb.ts`, add `@deprecated` mark to `getProcessingStatus` method from `db.ts`.
+- Support endpoints for retrieving and deleting existing invitations to a kb
 
 # 1.11.2 (2024-01-25)